Remove minify to make the code debuggable
I just got an error somewhere deeply inside of this library and cannot
figure out what's going on because the code here is minified and chrome
is not able to apply formatter.

Would be good to publish the new patch soon so I could find the problem
in my code.